requirements.txt
asgi-redis==0.14.1
channels==0.17.2
Cython==0.24.1
daphne==0.15.0
Django==1.10.1
h5py==2.6.0
matplotlib==1.5.3
mysql-python==1.2.5
nltk==3.2.1
numpy==1.11.1
pika==0.10.0
Pillow==3.3.0
pyyaml==3.12
scikit-image==0.12.3
scipy==0.17.1
ujson==1.35
uwsgi==2.0.13.1
websocket-client==0.37.0
